Move ppp and internal-stack to separate modules
